fix null filtering logic for IN list
