+++ pkg/planner/util/BUILD.bazel
@@ -19,6 +19,7 @@ go_library(
         "//pkg/parser/model",
         "//pkg/parser/mysql",
         "//pkg/planner/context",
+        "//pkg/planner/core/base",
         "//pkg/sessionctx/stmtctx",
         "//pkg/tablecodec",
         "//pkg/types",
