+++ pkg/planner/util/null_misc.go
@@ -18,76 +18,98 @@ import (
 	"github.com/pingcap/tidb/pkg/expression"
 	"github.com/pingcap/tidb/pkg/parser/ast"
 	"github.com/pingcap/tidb/pkg/planner/context"
+	"github.com/pingcap/tidb/pkg/planner/core/base"
 )
 
-// IsNullRejected check whether a condition is null-rejected
-// A condition would be null-rejected in one of following cases:
-// If it is a predicate containing a reference to an inner table that evaluates to UNKNOWN or FALSE
-// when one of its arguments is NULL.
-// If it is a conjunction containing a null-rejected condition as a conjunct.
-// If it is a disjunction of null-rejected conditions.
-func IsNullRejected(ctx context.PlanContext, schema *expression.Schema, expr expression.Expression) bool {
-	exprCtx := ctx.GetNullRejectCheckExprCtx()
-	expr = expression.PushDownNot(exprCtx, expr)
-	if expression.ContainOuterNot(expr) {
-		return false
+// allConstants checks if only the expression has only constants.
+func allConstants(ctx expression.BuildContext, expr expression.Expression) bool {
+	if expression.MaybeOverOptimized4PlanCache(ctx, []expression.Expression{expr}) {
+		return false // expression contains non-deterministic parameter
 	}
-	sc := ctx.GetSessionVars().StmtCtx
-	for _, cond := range expression.SplitCNFItems(expr) {
-		if isNullRejectedSpecially(ctx, schema, expr) {
-			return true
+	switch v := expr.(type) {
+	case *expression.ScalarFunction:
+		for _, arg := range v.GetArgs() {
+			if !allConstants(ctx, arg) {
+				return false
+			}
 		}
-
-		result := expression.EvaluateExprWithNull(exprCtx, schema, cond)
-		x, ok := result.(*expression.Constant)
-		if !ok {
-			continue
-		}
-		if x.Value.IsNull() {
-			return true
-		} else if isTrue, err := x.Value.ToBool(sc.TypeCtxOrDefault()); err == nil && isTrue == 0 {
-			return true
-		}
+		return true
+	case *expression.Constant:
+		return true
 	}
 	return false
 }
 
-// isNullRejectedSpecially handles some null-rejected cases specially, since the current in
-// EvaluateExprWithNull is too strict for some cases, e.g. #49616.
-func isNullRejectedSpecially(ctx context.PlanContext, schema *expression.Schema, expr expression.Expression) bool {
-	return specialNullRejectedCase1(ctx, schema, expr) // only 1 case now
+// isNullRejectedInList checks null filter for IN list using OR logic.
+// Reason is that null filtering through evaluation by isNullRejectedSimpleExpr
+// has problems with IN list.
+func isNullRejectedInList(ctx base.PlanContext, expr *expression.ScalarFunction, innerSchema *expression.Schema) bool {
+	for i, arg := range expr.GetArgs() {
+		if i > 0 {
+			newArgs := make([]expression.Expression, 0, 2)
+			newArgs = append(newArgs, expr.GetArgs()[0])
+			newArgs = append(newArgs, arg)
+			eQCondition, err := expression.NewFunction(ctx.GetExprCtx(), ast.EQ, expr.GetType(), newArgs...)
+			if err != nil {
+				return false
+			}
+			if !(isNullRejectedSimpleExpr(ctx, innerSchema, eQCondition)) {
+				return false
+			}
+		}
+	}
+	return true
 }
 
-// specialNullRejectedCase1 is mainly for #49616.
-// Case1 specially handles `null-rejected OR (null-rejected AND {others})`, then no matter what the result
-// of `{others}` is (True, False or Null), the result of this predicate is null, so this predicate is null-rejected.
-func specialNullRejectedCase1(ctx context.PlanContext, schema *expression.Schema, expr expression.Expression) bool {
-	isFunc := func(e expression.Expression, lowerFuncName string) *expression.ScalarFunction {
-		f, ok := e.(*expression.ScalarFunction)
-		if !ok {
-			return nil
-		}
-		if f.FuncName.L == lowerFuncName {
-			return f
-		}
-		return nil
-	}
-	orFunc := isFunc(expr, ast.LogicOr)
-	if orFunc == nil {
+// IsNullRejected takes care of complex predicates like this:
+// IsNullRejected(A OR B) = IsNullRejected(A) AND IsNullRejected(B)
+// IsNullRejected(A AND B) = IsNullRejected(A) OR IsNullRejected(B)
+func IsNullRejected(ctx base.PlanContext, innerSchema *expression.Schema, predicate expression.Expression) bool {
+	predicate = expression.PushDownNot(ctx.GetNullRejectCheckExprCtx(), predicate)
+	if expression.ContainOuterNot(predicate) {
 		return false
 	}
-	for i := 0; i < 2; i++ {
-		andFunc := isFunc(orFunc.GetArgs()[i], ast.LogicAnd)
-		if andFunc == nil {
-			continue
-		}
-		if !IsNullRejected(ctx, schema, orFunc.GetArgs()[1-i]) {
-			continue // the other side should be null-rejected: null-rejected OR (... AND ...)
-		}
-		for _, andItem := range expression.SplitCNFItems(andFunc) {
-			if IsNullRejected(ctx, schema, andItem) {
-				return true // hit the case in the comment: null-rejected OR (null-rejected AND ...)
+
+	switch expr := predicate.(type) {
+	case *expression.ScalarFunction:
+		if expr.FuncName.L == ast.LogicAnd {
+			if IsNullRejected(ctx, innerSchema, expr.GetArgs()[0]) {
+				return true
 			}
+			return IsNullRejected(ctx, innerSchema, expr.GetArgs()[1])
+		} else if expr.FuncName.L == ast.LogicOr {
+			if !(IsNullRejected(ctx, innerSchema, expr.GetArgs()[0])) {
+				return false
+			}
+			return IsNullRejected(ctx, innerSchema, expr.GetArgs()[1])
+		} else if expr.FuncName.L == ast.In {
+			return isNullRejectedInList(ctx, expr, innerSchema)
+		} else {
+			return isNullRejectedSimpleExpr(ctx, innerSchema, expr)
+		}
+	default:
+		return isNullRejectedSimpleExpr(ctx, innerSchema, predicate)
+	}
+}
+
+// isNullRejectedSimpleExpr check whether a condition is null-rejected
+// A condition would be null-rejected in one of following cases:
+// If it is a predicate containing a reference to an inner table that evaluates to UNKNOWN or FALSE
+// when one of its arguments is NULL.
+func isNullRejectedSimpleExpr(ctx context.PlanContext, schema *expression.Schema, expr expression.Expression) bool {
+	// The expression should reference at least one field in innerSchema or all constants.
+	if !expression.ExprReferenceSchema(expr, schema) && !allConstants(ctx.GetExprCtx(), expr) {
+		return false
+	}
+	exprCtx := ctx.GetNullRejectCheckExprCtx()
+	sc := ctx.GetSessionVars().StmtCtx
+	result := expression.EvaluateExprWithNull(exprCtx, schema, expr)
+	x, ok := result.(*expression.Constant)
+	if ok {
+		if x.Value.IsNull() {
+			return true
+		} else if isTrue, err := x.Value.ToBool(sc.TypeCtxOrDefault()); err == nil && isTrue == 0 {
+			return true
 		}
 	}
 	return false
